feat: 결제 없이 예약되도록 구현
