+++ app/src/main/java/com/kakaotech/team25/di/NetworkModule.kt
@@ -7,7 +7,6 @@ import com.kakaotech.team25.data.network.KakaoApi
 import com.kakaotech.team25.data.network.authenticator.HttpAuthenticator
 import com.kakaotech.team25.data.network.calladapter.ResultCallAdapter
 import com.kakaotech.team25.data.network.interceptor.TokenInterceptor
-import com.kakaotech.team25.data.network.remote.PaymentApiService
 import com.kakaotech.team25.data.network.services.RemoteSearchHospitalService
 import com.kakaotech.team25.data.network.typeadapter.LocalDateTypeAdapter
 import com.kakaotech.team25.data.remote.AccompanyApiService
@@ -94,10 +93,6 @@ object NetworkModule {
     @Singleton
     fun provideReservationApiService(@ServerRetrofit retrofit: Retrofit): ReservationApiService =
         retrofit.create(ReservationApiService::class.java)
-    @Provides
-    @Singleton
-    fun providePaymentApiService(@ServerRetrofit retrofit: Retrofit) : PaymentApiService =
-        retrofit.create(PaymentApiService::class.java)
 
     @Provides
     @Singleton
