--- app/src/main/java/com/kakaotech/team25/ui/reservation/PayStatus.kt
@@ -1,7 +1,0 @@
-package com.kakaotech.team25.ui.reservation
-
-enum class PayStatus {
-    DEFAULT,
-    SUCCESS,
-    FAILURE
-}
