--- app/src/main/res/layout/fragment_reservation_payment.xml
@@ -1,170 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <ImageView
-        android:id="@+id/back_btn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        android:padding="16dp"
-        android:src="@drawable/baseline_arrow_back_24"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="24dp"
-        android:text="@string/pay"
-        android:textSize="20sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/remove_card_text_view"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="20dp"
-        android:layout_marginBottom="10dp"
-        android:text="@string/remove_billing_key"
-        android:textColor="@color/blue_02"
-        android:textSize="16sp"
-        android:visibility="gone"
-        app:layout_constraintBottom_toTopOf="@id/info_payment_layout"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/info_payment_layout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="50dp"
-        android:layout_marginTop="50dp"
-        android:layout_marginEnd="50dp"
-        android:background="@drawable/payment_infor"
-        android:paddingBottom="20dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/back_btn">
-
-        <TextView
-            android:id="@+id/pay_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="20dp"
-            android:layout_marginTop="20dp"
-            android:text="@string/pay_info"
-            android:textSize="15sp"
-            android:textStyle="bold"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:id="@+id/deposit_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="15dp"
-            android:text="@string/basic_fee"
-            android:textSize="15sp"
-            app:layout_constraintStart_toStartOf="@id/pay_text"
-            app:layout_constraintTop_toBottomOf="@id/pay_text" />
-
-        <TextView
-            android:id="@+id/basic_fee_text_view"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginEnd="15dp"
-            android:text="@string/basic_usage_fee"
-            android:textSize="16sp"
-            app:layout_constraintEnd_toStartOf="@id/basic_won"
-            app:layout_constraintTop_toTopOf="@id/basic_won" />
-
-        <TextView
-            android:id="@+id/basic_won"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="150dp"
-            android:text="@string/won"
-            android:textSize="15sp"
-            app:layout_constraintBottom_toBottomOf="@+id/deposit_text"
-            app:layout_constraintStart_toEndOf="@id/deposit_text" />
-
-        <TextView
-            android:id="@+id/receipt_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="15dp"
-            android:text="@string/total_payment_amount"
-            android:textSize="15sp"
-            app:layout_constraintStart_toStartOf="@id/pay_text"
-            app:layout_constraintTop_toBottomOf="@id/deposit_text" />
-
-        <TextView
-            android:id="@+id/basic_fee"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginEnd="15dp"
-            android:text="@string/basic_usage_fee"
-            android:textSize="16sp"
-            android:textStyle="bold"
-            app:layout_constraintEnd_toStartOf="@id/basic_won"
-            app:layout_constraintTop_toTopOf="@id/basic_won" />
-
-        <TextView
-            android:id="@+id/total_fee_text_view"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginEnd="15dp"
-            android:text="@string/basic_usage_fee"
-            android:textColor="@color/blue"
-            android:textSize="16sp"
-            android:textStyle="bold"
-            app:layout_constraintEnd_toStartOf="@id/total_won_text_view"
-            app:layout_constraintTop_toTopOf="@id/total_won_text_view" />
-
-        <TextView
-            android:id="@+id/total_won_text_view"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="15dp"
-            android:text="@string/won"
-            android:textColor="@color/blue"
-            android:textSize="15sp"
-            app:layout_constraintStart_toStartOf="@+id/basic_won"
-            app:layout_constraintTop_toBottomOf="@id/basic_won" />
-
-        <TextView
-            android:layout_width="wrap_content"
-            app:layout_constraintTop_toBottomOf="@id/receipt_text"
-            android:layout_marginTop="20dp"
-            android:textColor="@color/red"
-            android:textStyle="bold"
-            android:textSize="13sp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            android:text="@string/additional_fees"
-            android:layout_height="wrap_content"/>
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <TextView
-        android:id="@+id/creditView"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:paddingTop="30dp"
-        android:paddingBottom="30dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        android:background="@drawable/credit_payment_background"
-        android:gravity="center"
-        android:text="@string/pay_credit_card"
-        android:textSize="18sp"
-        android:textColor="@color/white"
-        app:layout_constraintStart_toStartOf="@id/info_payment_layout"
-        app:layout_constraintEnd_toEndOf="@id/info_payment_layout"
-        app:layout_constraintTop_toBottomOf="@id/info_payment_layout" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
